setup renamed to setup_script.py
